requirements.txt
Django==5.2.2
djangorestframework==3.16.0
python-dotenv==1.1.0
paypalrestsdk==1.13.3
pytest-django==4.8.0
black==24.3.0
flake8==7.0.0 